+++ lua/codecompanion/strategies/chat/slash_commands/fetch.lua
@@ -9,7 +9,7 @@ local util_hash = require("codecompanion.utils.hash")
 
 local fmt = string.format
 
-CONSTANTS = {
+local CONSTANTS = {
   NAME = "Fetch",
   CACHE_PATH = vim.fn.stdpath("cache") .. "/codecompanion/urls",
 }
@@ -74,6 +74,8 @@ local function read_cache(chat, url, hash, opts)
   local p = Path:new(CONSTANTS.CACHE_PATH .. "/" .. hash)
   local cache = p:read()
 
+  log:debug("Fetch Slash Command: Restoring from cache for %s", url)
+
   return output(chat, {
     content = cache,
     url = url,
@@ -87,6 +89,7 @@ end
 local function write_cache(hash, data)
   local p = Path:new(CONSTANTS.CACHE_PATH .. "/" .. hash)
   p.filename = p:expand()
+  vim.fn.mkdir(CONSTANTS.CACHE_PATH, "p")
   p:touch({ parents = true })
   p:write(data or "", "w")
 end
@@ -103,6 +106,8 @@ local function fetch(chat, adapter, url, opts)
       return url
     end,
   }
+
+  log:debug("Fetch Slash Command: Fetching from %s", url)
 
   return client
     .new({
@@ -209,14 +214,27 @@ function SlashCommand:output(url, opts)
     return log:error("Failed to load the adapter for the fetch Slash Command")
   end
 
+  local function call_fetch()
+    return fetch(self.Chat, adapter, url, opts)
+  end
+
   local hash = util_hash.hash(url)
+
+  if opts and opts.ignore_cache then
+    log:debug("Fetch Slash Command: Ignoring cache")
+    return call_fetch()
+  end
+  if opts and opts.auto_restore_cache then
+    log:debug("Fetch Slash Command: Auto restoring from cache")
+    return read_cache(self.Chat, url, hash, opts)
+  end
 
   if is_cached(hash) then
     load_from_cache(self.Chat, url, hash, adapter, opts, function()
-      return fetch(self.Chat, adapter, url, opts)
+      return call_fetch()
     end)
   else
-    return fetch(self.Chat, adapter, url, opts)
+    return call_fetch()
   end
 end